Update to 7.0.7, update summary
Update summary, as it's confusing what this propduct is as there's also
seafile-server

here's some summary, what the products are:
- https://github.com/haiwen/seafile#source-repositories-for-seafile-components

Before version 6.0, the source code of "sync client daemon" and "server
core" are mixed together in haiwen/seafile. But after 6.0 version, the
server core is separated into its own repository.